Fix sepa payment detail page (#1579)
