+++ assets/css/admin.css
@@ -65,3 +65,8 @@
 .payment-method__brand--unknown {
 	background: no-repeat url( '../images/cards/unknown.svg' );
 }
+
+.payment-method__brand--sepa_debit {
+	background: no-repeat url( '../images/cards/sepa.svg' );
+	background-size: contain;
+}
